% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_status_prediction.R
\name{cell_status_prediction}
\alias{cell_status_prediction}
\title{cell status prediction}
\usage{
cell_status_prediction(svm_fit_cell, cell_var_list, peak_df)
}
\description{
cell status prediction
}
\examples{
cell_status_prediction()
}
