% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_data_with_status.R
\name{create_test_data_with_status}
\alias{create_test_data_with_status}
\title{Create test data with cell status from analytical method}
\usage{
create_test_data_with_status(fluo4_data_test, cell_status_df)
}
\description{
Create test data with cell status from analytical method
}
\examples{
create_test_data_with_status()
}
