% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rid_single_peak_cell.R
\name{rid_single_peak_cell}
\alias{rid_single_peak_cell}
\title{Single-peak cell elimination}
\usage{
rid_single_peak_cell(fluo4_data_test, single_peak_cells)
}
\description{
Single-peak cell elimination
}
\examples{
rid_single_peak_cell()
}
