% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rid_single_peak_signal.R
\name{rid_single_peak_signal}
\alias{rid_single_peak_signal}
\title{# Single-peak signal elimination}
\usage{
rid_single_peak_signal(peak_df, single_peak_cells)
}
\description{
# Single-peak signal elimination
}
\examples{
rid_single_peak_signal()
}
