% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_objects.R
\docType{data}
\name{svm_fit_peak}
\alias{svm_fit_peak}
\title{loading objects}
\format{
An object of class \code{svm.formula} (inherits from \code{svm}) of length 30.
}
\usage{
svm_fit_peak
}
\description{
loading objects
}
\keyword{datasets}
