% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMCaT.R
\docType{package}
\name{SVMCaT}
\alias{SVMCaT}
\title{SVMCaT Package}
\description{
Provides functions for high troughput machine learning algorithm-based assessment of calcium transient signal normality.
}
\author{
Hyun Hwang \email{hyun.hwang2@emory.edu}, Jingjing Yang \email{jingjing.yang@emory.edu}
}
